various small things. type cleanups, loop over the burst point to make sure it is at the right point. kick off the YP 'add' 5 seconds after source startup so that any stats are processed.
svn path=/icecast/trunk/icecast/; revision=9314

+++ src/connection.c
@@ -926,7 +926,7 @@ static void _handle_get_request (client_t *client, char *passed_uri)
          * the client, also.
          */
         if (source->max_listeners != -1 && 
-                source->listeners >= source->max_listeners) 
+                source->listeners >= (unsigned long)source->max_listeners) 
         {
             global_unlock();
             avl_tree_unlock(global.source_tree);
